fix multi disc

+++ mod/media/project_eris/etc/project_eris/SUP/emulationstation/playlist_es.sh
@@ -23,14 +23,12 @@ rom_dir="/media/roms/psx"
 lower_dir="${tmp_dir}${rom_dir}"
 original_m3u="/media/project_eris/etc/project_eris/SUP/retroarch"
 psc_region="$(cat "${original_games}/geninfo/REGION")"
-pal_game_ids=(SLUS-00594 SLUS-00776 SCUS-94163 SCUS-94164 SCUS-94165)
-ntscj_game_ids=(SLPS-01230 SLPS-01231 SLPS-01057 SLPS-01058 SLPS-01059 SLPM-86114 SLPM-86115)
+game_ids=(SLUS-00594 SLUS-00776 SCUS-94163 SCUS-94164 SCUS-94165 SLPS-01230 SLPS-01231 SLPS-01057 SLPS-01058 SLPS-01059 SLPM-86114 SLPM-86115)
+
 western_psc=(EK UC)
 
 # check if overlay is mounted and unmount it
 mount | grep -qs "${rom_dir}" && umount "${rom_dir}"
-
-# fix this
 find ${rom_dir} -type l -exec rm {} \;
 [ -d "${lower_dir}" ] && rm -rf "${lower_dir}"
 
@@ -39,42 +37,35 @@ find ${rom_dir} -type l -exec rm {} \;
 [ -f "${tmp_dir}/psxCleanup" ] && rm -rf "${tmp_dir}/psxCleanup"
 
 # find all PSX games and link them to ES rom folder
-find "$original_games" "$custom_games" -maxdepth 2 -regex ".*\.\(cue\|m3u\|bin\|pbp\|chd\|iso\|cbn\|ccd\|img\|mdf\|toc\|z\|znx\)" | while IFS= read -r game_file; do
+find "$original_games" "$custom_games" -maxdepth 2 -regex ".*\.\(cue\|m3u\|bin\|pbp\|chd\|iso\)" | while IFS= read -r game_file; do
   file_name="${game_file##*/}"
   game_title=${file_name%.*}
   ext=${game_file##*.}
   game_path="${game_file%/*}"
+  initial_cleanup="${game_title%%(*}"
+  disc_rm="${initial_cleanup%%Disc*}"
 
-  if [ ! -f "${lower_dir}/${file_name}" ]; then
-    ln -s "$game_file" "$lower_dir"
-  fi
+  [ ! -f "${lower_dir}/${file_name}" ] && ln -s "$game_file" "$lower_dir"
+
 
   if [ "${ext}" == "m3u" ]; then
-    echo "${game_title}" >>"${tmp_dir}/psxCleanup"
-    find "${game_path}" -name "*.cue" >"${lower_dir}/${file_name}"
+    echo "${disc_rm}" >>"${tmp_dir}/psxCleanup"
+    rm "${lower_dir}/${file_name}"
+    find "${game_path}" -name "*.cue" | while IFS= read -r cue_file; do
+      echo "../../games/${game_path##*/}/${cue_file##*/}" >>"${lower_dir}/${file_name}"
+    done
   fi
 done
-# DefKorns Only
-#     echo "Resident Evil 2 (Europe)
-# Parasite Eve (USA)
-# Parasite Eve II (USA)
-# Policenauts (Japan)
-# Heart of Darkness (USA)
-# Final Fantasy IX
-# Final Fantasy VIII (USA)
-# Chrono_Cross" >>"${tmp_dir}/psxCleanup"
 
 # Load playlists for multidisc original games
+for g in "${game_ids[@]}"; do
+  find "$lower_dir" -name "${g}.*" -exec rm {} +
+done
+
 if [[ "${western_psc[@]}" =~ "$psc_region" ]]; then
-  for g in "${pal_game_ids[@]}"; do
-    find "$lower_dir" -name "${g}.*" -exec rm {} +
-  done
   ln -s "${original_m3u}/SLUS-00594.m3u" "${lower_dir}"
   ln -s "${original_m3u}/SCUS-94163.m3u" "${lower_dir}"
 else
-  for g in "${ntscj_game_ids[@]}"; do
-    find "$lower_dir" -name "${g}.*" -exec rm {} +
-  done
   ln -s "${original_m3u}/SLPS-01230.m3u" "${lower_dir}"
   ln -s "${original_m3u}/SLPS-01057.m3u" "${lower_dir}"
   ln -s "${original_m3u}/SLPM-86114.m3u" "${lower_dir}"
